Grouped user stories into endpoint based story buckets
