+++ app/src/main/java/com/microsoft/office365/snippetapp/ActiveDirectoryStories/GetTenantDetailsStory.java
@@ -1,9 +1,10 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.ActiveDirectoryStories;
 
 import com.microsoft.directoryservices.TenantDetail;
+import com.microsoft.office365.snippetapp.helpers.BaseUserStory;
 import com.microsoft.office365.snippetapp.Snippets.UsersAndGroupsSnippets;
 import com.microsoft.office365.snippetapp.helpers.AuthenticationController;
 import com.microsoft.office365.snippetapp.helpers.Constants;
