+++ app/src/main/java/com/microsoft/office365/snippetapp/CalendarStories/CreateRecurringEventStory.java
@@ -1,100 +1,62 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.CalendarStories;
 
 import android.util.Log;
 
+import com.microsoft.office365.snippetapp.helpers.BaseUserStory;
 import com.microsoft.office365.snippetapp.R;
 import com.microsoft.office365.snippetapp.Snippets.CalendarSnippets;
 import com.microsoft.office365.snippetapp.helpers.APIErrorMessageHelper;
 import com.microsoft.office365.snippetapp.helpers.AuthenticationController;
 import com.microsoft.office365.snippetapp.helpers.GlobalValues;
-import com.microsoft.office365.snippetapp.helpers.StoryAction;
 import com.microsoft.office365.snippetapp.helpers.StoryResultFormatter;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
-//This story handles both of the following stories that appear in the UI list
-// based on strings passed in the constructor...
-//- Create an event (which is then deleted for cleanup)
-//- Delete an event (which is created first and then deleted)
-public class CreateOrDeleteEventStory extends BaseUserStory {
-    private final String CREATE_DESCRIPTION = "Adds a new calendar event";
-    private final String CREATE_TAG = "Create event story";
-    private final String CREATE_SUCCESS = "CreateEventStory: Event created.";
-    private final String CREATE_ERROR = "Create event exception: ";
-    private final String DELETE_DESCRIPTION = "Deletes a calendar event";
-    private final String DELETE_TAG = "Delete event story";
-    private final String DELETE_SUCCESS = "DeleteEventStory: Event deleted.";
-    private final String DELETE_ERROR = "Delete event exception: ";
-
-    private String mDescription;
-    private String mLogTag;
-    private String mSuccessDescription;
-    private String mErrorDescription;
-
-    public CreateOrDeleteEventStory(StoryAction action) {
-        switch (action) {
-            case CREATE: {
-                mDescription = CREATE_DESCRIPTION;
-                mLogTag = CREATE_TAG;
-                mSuccessDescription = CREATE_SUCCESS;
-                mErrorDescription = CREATE_ERROR;
-                break;
-            }
-            case DELETE: {
-                mDescription = DELETE_DESCRIPTION;
-                mLogTag = DELETE_TAG;
-                mSuccessDescription = DELETE_SUCCESS;
-                mErrorDescription = DELETE_ERROR;
-                break;
-            }
-        }
-    }
-
+public class CreateRecurringEventStory extends BaseUserStory {
     @Override
     public String execute() {
-        //PREPARE
+        boolean isStoryComplete;
+        String resultMessage;
+
         AuthenticationController
                 .getInstance()
                 .setResourceId(
-                        super.getO365MailResourceId());
+                        getO365MailResourceId());
 
         CalendarSnippets calendarSnippets = new CalendarSnippets(
                 getO365MailClient());
+        List<String> attendeeEmailAdresses = new ArrayList<>();
+        attendeeEmailAdresses.add(GlobalValues.USER_EMAIL);
 
-        List<String> attendeeEmailAddresses = new ArrayList<>();
-        attendeeEmailAddresses.add(GlobalValues.USER_EMAIL);
-        String newEventId = "";
-        //ACT
         try {
-            newEventId = calendarSnippets.createCalendarEvent(
+            //Create a recurring event
+            String newEventId = calendarSnippets.createRecurringCalendarEvent(
                     getStringResource(R.string.calendar_subject_text)
                     , getStringResource(R.string.calendar_body_text)
-                    , java.util.Calendar.getInstance()
-                    , java.util.Calendar.getInstance()
-                    , attendeeEmailAddresses);
+                    , attendeeEmailAdresses);
 
-            //Delete event
+            //Cleanup by deleting the event
             calendarSnippets.deleteCalendarEvent(newEventId);
+            isStoryComplete = true;
+            resultMessage = "Create recurring event: Recurring event created";
         } catch (ExecutionException | InterruptedException e) {
-            e.printStackTrace();
+            isStoryComplete = false;
             String formattedException = APIErrorMessageHelper.getErrorMessage(e.getMessage());
-            Log.e(mLogTag, formattedException);
-            return StoryResultFormatter.wrapResult(
-                    mErrorDescription + formattedException
-                    , false
-            );
+            Log.e("CreateRecurringEvent", formattedException);
+            resultMessage = "Create recurring event exception: " + formattedException;
         }
-        return StoryResultFormatter.wrapResult(mSuccessDescription, true);
+
+        return StoryResultFormatter.wrapResult(resultMessage, isStoryComplete);
     }
 
     @Override
     public String getDescription() {
-        return mDescription;
+        return "Create a recurring event";
     }
 }
 // *********************************************************
