+++ app/src/main/java/com/microsoft/office365/snippetapp/ContactStories/CreateOrDeleteContactStory.java
@@ -1,73 +1,99 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.ContactStories;
 
 import android.util.Log;
 
+import com.microsoft.office365.snippetapp.helpers.BaseUserStory;
 import com.microsoft.office365.snippetapp.R;
 import com.microsoft.office365.snippetapp.Snippets.ContactsSnippets;
 import com.microsoft.office365.snippetapp.helpers.APIErrorMessageHelper;
 import com.microsoft.office365.snippetapp.helpers.AuthenticationController;
+import com.microsoft.office365.snippetapp.helpers.StoryAction;
 import com.microsoft.office365.snippetapp.helpers.StoryResultFormatter;
-import com.microsoft.outlookservices.Contact;
 
-import java.util.List;
 import java.util.concurrent.ExecutionException;
 
-public class GetFilteredContactsWithSurnameStory extends BaseUserStory {
+//This story handles both of the following stories that appear in the UI list
+// based on strings passed in the constructor...
+//- Create a contact (which is then deleted for cleanup)
+//- Delete a contact (which is created first and then deleted)
+public class CreateOrDeleteContactStory extends BaseUserStory {
+    private final String CREATE_DESCRIPTION = "Creates a new contact";
+    private final String CREATE_TAG = "Create contact story";
+    private final String CREATE_SUCCESS = "CreateContactStory: Contact created.";
+    private final String CREATE_ERROR = "Create contact exception: ";
+    private final String DELETE_DESCRIPTION = "Deletes a contact";
+    private final String DELETE_TAG = "Delete contact story";
+    private final String DELETE_SUCCESS = "DeleteContactStory: Contact deleted.";
+    private final String DELETE_ERROR = "Delete contact exception: ";
+
+    private String mDescription;
+    private String mLogTag;
+    private String mSuccessDescription;
+    private String mErrorDescription;
+
+    public CreateOrDeleteContactStory(StoryAction action) {
+        switch (action) {
+            case CREATE: {
+                mDescription = CREATE_DESCRIPTION;
+                mLogTag = CREATE_TAG;
+                mSuccessDescription = CREATE_SUCCESS;
+                mErrorDescription = CREATE_ERROR;
+                break;
+            }
+            case DELETE: {
+                mDescription = DELETE_DESCRIPTION;
+                mLogTag = DELETE_TAG;
+                mSuccessDescription = DELETE_SUCCESS;
+                mErrorDescription = DELETE_ERROR;
+                break;
+            }
+        }
+    }
 
     @Override
     public String execute() {
-        boolean isStoryComplete = false;
-        StringBuilder storyResultText = new StringBuilder("FilterContactsBySurnameStory: ");
-
-        String surname = getStringResource(R.string.contacts_last_name);
+        String returnValue = StoryResultFormatter.wrapResult(
+                mLogTag, false
+        );
         AuthenticationController
                 .getInstance()
-                .setResourceId(getO365MailResourceId());
-        ContactsSnippets contactsSnippets = new ContactsSnippets(getO365MailClient());
+                .setResourceId(
+                        getO365MailResourceId());
 
+        ContactsSnippets contactsSnippets = new ContactsSnippets(getO365MailClient());
         try {
-            //Create a contact that we can test against
             String contactId = contactsSnippets.createContact(
                     getStringResource(R.string.contacts_email),
                     getStringResource(R.string.contacts_business_phone),
                     getStringResource(R.string.contacts_home_phone),
                     getStringResource(R.string.contacts_first_name),
-                    surname);
-
-            //Find the new test contact
-            List<Contact> contacts = contactsSnippets.getContactsWithSurname(surname);
-            for (Contact contact : contacts) {
-                if (contact.getSurname().equals(surname)) {
-                    break;
-                }
+                    getStringResource(R.string.contacts_last_name));
+            if (contactId.length() > 0) {
+                contactsSnippets.deleteContact(contactId);
+                return StoryResultFormatter.wrapResult(
+                        mSuccessDescription, true
+                );
             }
-            //Delete the test contact from tenant
-            contactsSnippets.deleteContact(contactId);
-
-            //Story is completed
-            isStoryComplete = true;
-            storyResultText.append("Contact with surname ")
-                    .append(surname)
-                    .append(" found.");
         } catch (ExecutionException | InterruptedException e) {
             e.printStackTrace();
             String formattedException = APIErrorMessageHelper.getErrorMessage(e.getMessage());
-            Log.e("ContactFilter", formattedException);
-            storyResultText.append("Filter contacts by surname exception: ")
-                    .append(formattedException);
-            isStoryComplete = false;
+            Log.e(mLogTag, formattedException);
+            return StoryResultFormatter.wrapResult(
+                    mErrorDescription + formattedException
+                    , false
+            );
+
         }
-        return StoryResultFormatter.wrapResult(storyResultText.toString(), isStoryComplete);
+        return returnValue;
     }
 
     @Override
     public String getDescription() {
-        return "Gets contacts filtered by surname";
+        return mDescription;
     }
-
 
 }
 // *********************************************************
