+++ app/src/main/java/com/microsoft/office365/snippetapp/EmailStories/BaseEmailUserStory.java
@@ -1,8 +1,9 @@
 /*
 *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
 */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.EmailStories;
 
+import com.microsoft.office365.snippetapp.helpers.BaseUserStory;
 import com.microsoft.office365.snippetapp.Snippets.EmailSnippets;
 import com.microsoft.outlookservices.Message;
