+++ app/src/main/java/com/microsoft/office365/snippetapp/EmailStories/ForwardEmailMessageStory.java
@@ -1,10 +1,9 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.EmailStories;
 
-import android.util.Log;
-
+import com.microsoft.office365.snippetapp.EmailStories.BaseEmailUserStory;
 import com.microsoft.office365.snippetapp.R;
 import com.microsoft.office365.snippetapp.Snippets.EmailSnippets;
 import com.microsoft.office365.snippetapp.helpers.APIErrorMessageHelper;
@@ -15,22 +14,18 @@ import com.microsoft.outlookservices.Message;
 
 import java.util.Date;
 
-public class SendEmailWithMessageAttachStory extends BaseEmailUserStory {
-
-    public static final String STORY_DESCRIPTION = "Sends an email message with a message attachment";
-    public static final String SENT_NOTICE = "Email sent with subject line:";
-    public static final boolean IS_INLINE = false;
-
+public class ForwardEmailMessageStory extends BaseEmailUserStory {
 
     @Override
     public String execute() {
         String returnResult = "";
-        try {
-            AuthenticationController
-                    .getInstance()
-                    .setResourceId(
-                            getO365MailResourceId());
 
+        AuthenticationController
+                .getInstance()
+                .setResourceId(
+                        getO365MailResourceId());
+
+        try {
             EmailSnippets emailSnippets = new EmailSnippets(
                     getO365MailClient());
 
@@ -43,56 +38,35 @@ public class SendEmailWithMessageAttachStory extends BaseEmailUserStory {
                     , getStringResource(R.string.mail_subject_text)
                             + uniqueGUID, getStringResource(R.string.mail_body_text));
 
-
-            Message messageToAttach = GetAMessageFromEmailFolder(emailSnippets,
+            //Get the new message
+            Message messageToForward = GetAMessageFromEmailFolder(emailSnippets,
                     getStringResource(R.string.mail_subject_text)
                             + uniqueGUID, getStringResource(R.string.Email_Folder_Inbox));
 
-            if (messageToAttach != null) {
-                //Create a new email message but do not send yet
-                String newEmailId = emailSnippets.addDraftMail(
-                        GlobalValues.USER_EMAIL
-                        , getStringResource(R.string.mail_subject_text) + uniqueGUID
-                        , getStringResource(R.string.mail_body_text));
-
-                //Attach email message to new draft email
-                emailSnippets.addItemAttachment(
-                        newEmailId
-                        , messageToAttach
-                        , IS_INLINE);
-
-                //Send draft email
-                emailSnippets.sendMail(newEmailId);
-
-                DeleteAMessageFromMailFolder(emailSnippets,
-                        getStringResource(R.string.mail_subject_text)
-                                + uniqueGUID,
-                        getStringResource(R.string.Email_Folder_Sent));
-
-                returnResult = StoryResultFormatter.wrapResult(
-                        STORY_DESCRIPTION, true
-                );
+            String forwardEmailId = emailSnippets.forwardMail(messageToForward.getId());
+            //3. Delete the email using the ID
+            emailSnippets.deleteMail(messageToForward.getId());
+            if (forwardEmailId.length() > 0) {
+                emailSnippets.deleteMail(forwardEmailId);
             }
 
-
+            return StoryResultFormatter.wrapResult(
+                    "Forward email message story: ", true
+            );
         } catch (Exception ex) {
             String formattedException = APIErrorMessageHelper.getErrorMessage(ex.getMessage());
-            Log.e("Send msg w/ message ", formattedException);
-            returnResult = StoryResultFormatter.wrapResult(
-                    "Send mail exception: "
-                            + formattedException
-                    , false
+            return StoryResultFormatter.wrapResult(
+                    "Forward email message story: " + formattedException, false
             );
         }
-        return returnResult;
     }
 
     @Override
     public String getDescription() {
-        return STORY_DESCRIPTION;
+        return "Forward an email message";
     }
-}
 
+}
 // *********************************************************
 //
 // O365-Android-Snippets, https://github.com/OfficeDev/O365-Android-Snippets
