+++ app/src/main/java/com/microsoft/office365/snippetapp/EmailStories/GetEmailAttachmentsStory.java
@@ -1,7 +1,7 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.EmailStories;
 
 import android.util.Log;
 
@@ -11,12 +11,18 @@ import com.microsoft.office365.snippetapp.helpers.APIErrorMessageHelper;
 import com.microsoft.office365.snippetapp.helpers.AuthenticationController;
 import com.microsoft.office365.snippetapp.helpers.GlobalValues;
 import com.microsoft.office365.snippetapp.helpers.StoryResultFormatter;
+import com.microsoft.outlookservices.Attachment;
+import com.microsoft.outlookservices.FileAttachment;
+import com.microsoft.outlookservices.Message;
+
+import java.io.UnsupportedEncodingException;
+import java.util.Date;
+import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 
-public class SendEmailWithTextFileAttachmentStory extends BaseEmailUserStory {
-
-    public static final String STORY_DESCRIPTION = "Sends an email message with a text file attachment";
-    public static final String SENT_NOTICE = "Email sent with subject line:";
+public class GetEmailAttachmentsStory extends BaseEmailUserStory {
+    public static final String SENT_NOTICE = "Attachment email sent with subject line:";
     public static final boolean IS_INLINE = false;
 
     @Override
@@ -30,13 +36,13 @@ public class SendEmailWithTextFileAttachmentStory extends BaseEmailUserStory {
 
             EmailSnippets emailSnippets = new EmailSnippets(
                     getO365MailClient());
-
             //1. Send an email and store the ID
             String uniqueGUID = java.util.UUID.randomUUID().toString();
+            String mailSubject = getStringResource(R.string.mail_subject_text) + uniqueGUID;
 
             //Add a new email to the user's draft folder
             String emailID = emailSnippets.addDraftMail(GlobalValues.USER_EMAIL,
-                    getStringResource(R.string.mail_subject_text) + uniqueGUID,
+                    mailSubject,
                     getStringResource(R.string.mail_body_text));
 
             //Add a text file attachment to the mail added to the draft folder
@@ -47,25 +53,52 @@ public class SendEmailWithTextFileAttachmentStory extends BaseEmailUserStory {
 
             String draftMessageID = emailSnippets.getMailMessageById(emailID).getId();
 
-            //Send the mail with attachments
-            //build string for test results on UI
-            StringBuilder sb = new StringBuilder();
-            sb.append(SENT_NOTICE);
-            sb.append(getStringResource(R.string.mail_subject_text) + uniqueGUID);
-            returnResult = StoryResultFormatter.wrapResult(sb.toString(), true);
-
+            //UTC time Immediately before message is sent
+            Date sendDate = new Date();
             //Send the draft email to the recipient
             emailSnippets.sendMail(draftMessageID);
 
-        } catch (Exception ex) {
+            //Get the new message
+            Message sentMessage = GetAMessageFromEmailFolder(emailSnippets,
+                    getStringResource(R.string.mail_subject_text)
+                            + uniqueGUID, getStringResource(R.string.Email_Folder_Inbox));
+
+
+            StringBuilder sb = new StringBuilder();
+            sb.append(SENT_NOTICE);
+            sb.append(getStringResource(R.string.mail_subject_text) + uniqueGUID);
+            if (sentMessage.getId().length() > 0) {
+                List<Attachment> attachments = emailSnippets.getAttachmentsFromEmailMessage(
+                        sentMessage.getId());
+                //Send the mail with attachments
+                //build string for test results on UI
+                for (Attachment attachment : attachments) {
+                    if (attachment instanceof FileAttachment) {
+                        FileAttachment fileAttachment = (FileAttachment) attachment;
+                        String fileContents = new String(fileAttachment.getContentBytes(), "UTF-8");
+                        sb.append(fileContents);
+                        sb.append("/n");
+                    }
+                }
+                returnResult = StoryResultFormatter.wrapResult(sb.toString(), true);
+            } else
+                returnResult = StoryResultFormatter.wrapResult(sb.toString(), false);
+
+
+            //3. Delete the email using the ID
+            // Boolean result = emailSnippets.deleteMail(emailID);
+
+        } catch (ExecutionException | InterruptedException ex) {
             String formattedException = APIErrorMessageHelper.getErrorMessage(ex.getMessage());
-            Log.e("Send email story", formattedException);
+            Log.e("GetEmailAttachments", formattedException);
             return StoryResultFormatter.wrapResult(
-                    "Send mail exception: "
+                    "Get email attachments exception: "
                             + formattedException
                     , false
             );
 
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
         }
         return returnResult;
 
@@ -73,7 +106,7 @@ public class SendEmailWithTextFileAttachmentStory extends BaseEmailUserStory {
 
     @Override
     public String getDescription() {
-        return STORY_DESCRIPTION;
+        return "Gets the attachments from an email message";
     }
 }
 // *********************************************************
