+++ app/src/main/java/com/microsoft/office365/snippetapp/EmailStories/ReplyToEmailMessageStory.java
@@ -1,9 +1,7 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
-
-import android.util.Log;
+package com.microsoft.office365.snippetapp.EmailStories;
 
 import com.microsoft.office365.snippetapp.R;
 import com.microsoft.office365.snippetapp.Snippets.EmailSnippets;
@@ -13,56 +11,67 @@ import com.microsoft.office365.snippetapp.helpers.GlobalValues;
 import com.microsoft.office365.snippetapp.helpers.StoryResultFormatter;
 import com.microsoft.outlookservices.Message;
 
-public class SendEmailMessageStory extends BaseEmailUserStory {
+import java.util.Date;
+import java.util.concurrent.ExecutionException;
+
+//Create a new email, send to yourself, reply to the email, and delete sent mail
+public class ReplyToEmailMessageStory extends BaseEmailUserStory {
+
 
     @Override
     public String execute() {
-        String returnResult = "";
+
+        AuthenticationController
+                .getInstance()
+                .setResourceId(
+                        getO365MailResourceId());
+
         try {
-
-            AuthenticationController
-                    .getInstance()
-                    .setResourceId(
-                            getO365MailResourceId());
-
             EmailSnippets emailSnippets = new EmailSnippets(
                     getO365MailClient());
 
+            //Store the date and time that the email is sent in UTC
+            Date sentDate = new Date();
             //1. Send an email and store the ID
             String uniqueGUID = java.util.UUID.randomUUID().toString();
-            String subject = getStringResource(R.string.mail_subject_text) + uniqueGUID;
-            emailSnippets.createAndSendMail(GlobalValues.USER_EMAIL,
-                    subject,
-                    getStringResource(R.string.mail_body_text));
+            String emailID = emailSnippets.createAndSendMail(
+                    GlobalValues.USER_EMAIL
+                    , getStringResource(R.string.mail_subject_text)
+                            + uniqueGUID
+                    , getStringResource(R.string.mail_body_text));
 
-            //3. Delete the email using the ID
-            Message message = GetAMessageFromEmailFolder(emailSnippets, subject, getStringResource(R.string.Email_Folder_Inbox));
-            emailSnippets.deleteMail(message.getId());
+            //Get the new message
+            Message message = GetAMessageFromEmailFolder(emailSnippets,
+                    getStringResource(R.string.mail_subject_text)
+                            + uniqueGUID, getStringResource(R.string.Email_Folder_Inbox));
 
-            //build string for test results on UI
-            StringBuilder sb = new StringBuilder();
-            sb.append("Email is added");
-            returnResult = StoryResultFormatter.wrapResult(sb.toString(), true);
-        } catch (Exception ex) {
+            if (message.getId().length() > 0) {
+                String replyEmailId = emailSnippets.replyToEmailMessage(
+                        message.getId()
+                        , getStringResource(R.string.mail_body_text));
+                //3. Delete the email using the ID
+                emailSnippets.deleteMail(message.getId());
+                if (replyEmailId.length() > 0) {
+                    emailSnippets.deleteMail(replyEmailId);
+                }
+                return StoryResultFormatter.wrapResult(
+                        "Reply to email message story: ", true);
+            } else {
+                return StoryResultFormatter.wrapResult(
+                        "Reply to email message story: ", false);
+            }
+        } catch (ExecutionException | InterruptedException ex) {
             String formattedException = APIErrorMessageHelper.getErrorMessage(ex.getMessage());
-            Log.e("Send email story", formattedException);
             return StoryResultFormatter.wrapResult(
-                    "Send mail exception: "
-                            + formattedException
-                    , false
+                    "Reply to email message story: " + formattedException, false
             );
-
         }
-        return returnResult;
     }
 
     @Override
     public String getDescription() {
-
-        return "Sends an email message";
+        return "Reply to an email message";
     }
-
-
 }
 // *********************************************************
 //
