+++ app/src/main/java/com/microsoft/office365/snippetapp/EmailStories/SendEmailMessageStory.java
@@ -1,7 +1,9 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.EmailStories;
+
+import android.util.Log;
 
 import com.microsoft.office365.snippetapp.R;
 import com.microsoft.office365.snippetapp.Snippets.EmailSnippets;
@@ -11,59 +13,55 @@ import com.microsoft.office365.snippetapp.helpers.GlobalValues;
 import com.microsoft.office365.snippetapp.helpers.StoryResultFormatter;
 import com.microsoft.outlookservices.Message;
 
-import java.util.Date;
-
-public class ForwardEmailMessageStory extends BaseEmailUserStory {
+public class SendEmailMessageStory extends BaseEmailUserStory {
 
     @Override
     public String execute() {
         String returnResult = "";
-
-        AuthenticationController
-                .getInstance()
-                .setResourceId(
-                        getO365MailResourceId());
-
         try {
+
+            AuthenticationController
+                    .getInstance()
+                    .setResourceId(
+                            getO365MailResourceId());
+
             EmailSnippets emailSnippets = new EmailSnippets(
                     getO365MailClient());
 
-            //Store the date and time that the email is sent in UTC
-            Date sentDate = new Date();
             //1. Send an email and store the ID
             String uniqueGUID = java.util.UUID.randomUUID().toString();
-            String emailID = emailSnippets.createAndSendMail(
-                    GlobalValues.USER_EMAIL
-                    , getStringResource(R.string.mail_subject_text)
-                            + uniqueGUID, getStringResource(R.string.mail_body_text));
+            String subject = getStringResource(R.string.mail_subject_text) + uniqueGUID;
+            emailSnippets.createAndSendMail(GlobalValues.USER_EMAIL,
+                    subject,
+                    getStringResource(R.string.mail_body_text));
 
-            //Get the new message
-            Message messageToForward = GetAMessageFromEmailFolder(emailSnippets,
-                    getStringResource(R.string.mail_subject_text)
-                            + uniqueGUID, getStringResource(R.string.Email_Folder_Inbox));
-
-            String forwardEmailId = emailSnippets.forwardMail(messageToForward.getId());
             //3. Delete the email using the ID
-            emailSnippets.deleteMail(messageToForward.getId());
-            if (forwardEmailId.length() > 0) {
-                emailSnippets.deleteMail(forwardEmailId);
-            }
+            Message message = GetAMessageFromEmailFolder(emailSnippets, subject, getStringResource(R.string.Email_Folder_Inbox));
+            emailSnippets.deleteMail(message.getId());
 
-            return StoryResultFormatter.wrapResult(
-                    "Forward email message story: ", true
-            );
+            //build string for test results on UI
+            StringBuilder sb = new StringBuilder();
+            sb.append("Email is added");
+            returnResult = StoryResultFormatter.wrapResult(sb.toString(), true);
         } catch (Exception ex) {
             String formattedException = APIErrorMessageHelper.getErrorMessage(ex.getMessage());
+            Log.e("Send email story", formattedException);
             return StoryResultFormatter.wrapResult(
-                    "Forward email message story: " + formattedException, false
+                    "Send mail exception: "
+                            + formattedException
+                    , false
             );
+
         }
+        return returnResult;
     }
 
     @Override
     public String getDescription() {
-        return "Forward an email message";
+
+        return "Sends an email message";
     }
+
 
 }
 // *********************************************************
