+++ app/src/main/java/com/microsoft/office365/snippetapp/helpers/BaseUserStory.java
@@ -1,7 +1,7 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.snippetapp.O365Stories;
+package com.microsoft.office365.snippetapp.helpers;
 
 
 import android.content.res.AssetFileDescriptor;
