+++ app/src/main/java/com/microsoft/office365/snippetapp/helpers/StoryList.java
@@ -5,32 +5,31 @@ package com.microsoft.office365.snippetapp.helpers;
 
 import android.content.Context;
 
-import com.microsoft.office365.snippetapp.O365Stories.BaseUserStory;
-import com.microsoft.office365.snippetapp.O365Stories.CreateOrDeleteContactStory;
-import com.microsoft.office365.snippetapp.O365Stories.CreateOrDeleteEventStory;
-import com.microsoft.office365.snippetapp.O365Stories.CreateOrDeleteFileStory;
-import com.microsoft.office365.snippetapp.O365Stories.CreateOrDeleteOneDriveFolder;
-import com.microsoft.office365.snippetapp.O365Stories.CreateRecurringEventStory;
-import com.microsoft.office365.snippetapp.O365Stories.DownloadFileStory;
-import com.microsoft.office365.snippetapp.O365Stories.EventsFetcherStory;
-import com.microsoft.office365.snippetapp.O365Stories.ForwardEmailMessageStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetADGroupsStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetADUsersStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetContactsStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetEmailAttachmentsStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetEmailMessagesStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetFilesAndFoldersStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetFilteredContactsWithSurnameStory;
-import com.microsoft.office365.snippetapp.O365Stories.GetFilteredImportantEvents;
-import com.microsoft.office365.snippetapp.O365Stories.GetTenantDetailsStory;
-import com.microsoft.office365.snippetapp.O365Stories.ReplyToEmailMessageStory;
-import com.microsoft.office365.snippetapp.O365Stories.RespondToCalendarEventInviteStory;
-import com.microsoft.office365.snippetapp.O365Stories.SendEmailMessageStory;
-import com.microsoft.office365.snippetapp.O365Stories.SendEmailWithMessageAttachStory;
-import com.microsoft.office365.snippetapp.O365Stories.SendEmailWithTextFileAttachmentStory;
-import com.microsoft.office365.snippetapp.O365Stories.UpdateContactStory;
-import com.microsoft.office365.snippetapp.O365Stories.UpdateEventStory;
-import com.microsoft.office365.snippetapp.O365Stories.UpdateFileContentsOnServerStory;
+import com.microsoft.office365.snippetapp.ContactStories.CreateOrDeleteContactStory;
+import com.microsoft.office365.snippetapp.CalendarStories.CreateOrDeleteEventStory;
+import com.microsoft.office365.snippetapp.FileFolderStories.CreateOrDeleteFileStory;
+import com.microsoft.office365.snippetapp.FileFolderStories.CreateOrDeleteOneDriveFolder;
+import com.microsoft.office365.snippetapp.CalendarStories.CreateRecurringEventStory;
+import com.microsoft.office365.snippetapp.FileFolderStories.DownloadFileStory;
+import com.microsoft.office365.snippetapp.CalendarStories.EventsFetcherStory;
+import com.microsoft.office365.snippetapp.EmailStories.ForwardEmailMessageStory;
+import com.microsoft.office365.snippetapp.ActiveDirectoryStories.GetADGroupsStory;
+import com.microsoft.office365.snippetapp.ActiveDirectoryStories.GetADUsersStory;
+import com.microsoft.office365.snippetapp.ContactStories.GetContactsStory;
+import com.microsoft.office365.snippetapp.EmailStories.GetEmailAttachmentsStory;
+import com.microsoft.office365.snippetapp.EmailStories.GetEmailMessagesStory;
+import com.microsoft.office365.snippetapp.FileFolderStories.GetFilesAndFoldersStory;
+import com.microsoft.office365.snippetapp.ContactStories.GetFilteredContactsWithSurnameStory;
+import com.microsoft.office365.snippetapp.CalendarStories.GetFilteredImportantEvents;
+import com.microsoft.office365.snippetapp.ActiveDirectoryStories.GetTenantDetailsStory;
+import com.microsoft.office365.snippetapp.EmailStories.ReplyToEmailMessageStory;
+import com.microsoft.office365.snippetapp.CalendarStories.RespondToCalendarEventInviteStory;
+import com.microsoft.office365.snippetapp.EmailStories.SendEmailMessageStory;
+import com.microsoft.office365.snippetapp.EmailStories.SendEmailWithMessageAttachStory;
+import com.microsoft.office365.snippetapp.EmailStories.SendEmailWithTextFileAttachmentStory;
+import com.microsoft.office365.snippetapp.ContactStories.UpdateContactStory;
+import com.microsoft.office365.snippetapp.CalendarStories.UpdateEventStory;
+import com.microsoft.office365.snippetapp.FileFolderStories.UpdateFileContentsOnServerStory;
 
 import java.util.ArrayList;
 import java.util.Arrays;
